lock: require uv for uv.lock resolution; fail fast instead of falling back
Removes the built-in closure walker (resolveUV/runtimeClosure/ParseUVLockFile/
ParseAny) entirely so there is no second, potentially divergent resolver. uv.lock
resolution now goes only through `uv export`; if uv is not on PATH the build
fails with an actionable error pointing at requirements.txt + --find-links.
uv.lock parsing is kept solely to attach wheel/sdist URLs.
