+++ pymage/DESIGN.md
@@ -32,11 +32,12 @@ because it exploits content-addressed layering:
   pre-build a wheel out-of-band and supply it via `--find-links`.
 - Replacing dependency *resolution*. We delegate resolution to existing,
   correct tools and consume their output: for uv projects we shell out to
-  `uv export --frozen --no-dev` (when uv is available) to get the exact
-  per-target requirement set — closure, extras, groups, and workspace selection
-  are uv's responsibility, not ours. pymage only evaluates the markers uv emits
-  and maps each pin to a wheel. A built-in lock walker remains as a fallback for
-  environments without uv. Likewise, install layout is validated against a real
+  `uv export --frozen --no-dev` to get the exact per-target requirement set —
+  closure, extras, groups, and workspace selection are uv's responsibility, not
+  ours. pymage only evaluates the markers uv emits and maps each pin to a wheel.
+  uv is therefore **required** to build from a uv.lock; we fail fast if it's
+  absent rather than fall back to a second, divergent resolver. Likewise, install
+  layout is validated against a real
   `uv pip install` in a conformance test (installing a wheel runs no code, so
   there is no RCE risk in using it as an oracle).
 - A general-purpose Dockerfile interpreter. This is a focused Python app builder
